pyproject.toml
[tool.poetry]
name = "zoomcamp-mlops"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.1"
prefect = "^2.16.8"
mlflow = "^2.11.3"
ipython = "^8.23.0"
python-dotenv = "^1.0.1"
scikit-learn = "^1.4.1.post1"
httpx = "^0.27.0"
black = "^24.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
